Three-wheeled direction controlled robot using Gyroscope Sensor(MPU-6050) and Roboclaw motor controller.
Three-wheeled direction-controlled robot using a gyroscope sensor and P.I.D. to set the robot at the initial angle even after moving or changing the directions of the bot physically or externally. A RoboClaw motor controller is used for a three-wheeled drive with an Omnidirectional wheel to control the robot. An I2C communication protocol is used for receiving data from the Gyroscopic Sensor.